third_party/WebKit/LayoutTests/platform/win/external/wpt/webmessaging/postMessage_ArrayBuffer.sub-expected.txt

This is a testharness.js-based test.
PASS Test Description: postMessage with ArrayBuffer object. 
PASS ArrayBuffer is supported. 
Harness: the test ran to completion.
